requirements.txt
python-telegram-bot
pillow
python-dotenv